refactor: extract layout and not found components
