+++ src/app/app.tsx
@@ -1,20 +1,10 @@
 import './app.css';
 import SpeciesLookup from '../components/species-lookup.tsx';
 import ErrorBoundary from './error-boundary.tsx';
-import Header from './header.tsx';
-import AppCrashingButton from '../components/app-crashing-button.tsx';
 import MessageBox from '../components/message-box.tsx';
-import { NavLink, Outlet, Route, Routes } from 'react-router';
-
-function Layout() {
-  return (
-    <div className="container">
-      <Header />
-      <Outlet />
-      <AppCrashingButton />
-    </div>
-  );
-}
+import { Outlet, Route, Routes } from 'react-router';
+import NotFound from './not-found.tsx';
+import Layout from './layout.tsx';
 
 function SpecieDetail() {
   return (
@@ -25,27 +15,12 @@ function SpecieDetail() {
   );
 }
 
-function NotFound() {
-  return (
-    <MessageBox
-      message={
-        <>
-          NOT FOUND |{' '}
-          <NavLink to="/" className="link">
-            GO HOME
-          </NavLink>
-        </>
-      }
-    />
-  );
-}
-
 function App() {
   return (
     <Routes>
       <Route element={<Layout />}>
         <Route index element={<SpeciesLookup />} />
-        <Route path="/species" element={<SpeciesLookup />}>
+        <Route path="species" element={<SpeciesLookup />}>
           <Route path=":specieId" element={<SpecieDetail />} />
         </Route>
         <Route path="*" element={<NotFound />} />
